// frontend/src/generated/model/ingredient-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IngredientUnitDto } from './ingredient-unit-dto';


/**
 * DTO containing all ingredient information.
 */
export interface IngredientDto { 
    /**
     * Unique id of the ingredient.
     */
    id: number;
    /**
     * Name of the ingredient.
     */
    name: string;
    defaultUnit: IngredientUnitDto;
}